refactor(ui): drop the useVerifyWithZKPassport hook
The children render prop already gives a custom button
{ verify, isLoading, status, error }, so the hook was a second public
path to the same thing. The button now holds its own state over the
shared createVerification controller, like the DOM button does.
